lib: ICC compatibility workaround
Dummy FRR_MODULE_SETUP added to lib/libfrr.c to fix the build when
using the Intel C compiler. This is not a complete fix, as e.g. make
check is broken because of problems of FRR_MODULE_SETUP-related missing
symbols in the ISIS daemon (I've tried to add it a dummy declaration too,
which works for fixing the make check for ICC, but it generates
duplicated symbols for e.g. GCC).

Signed-off-by: F. Aragon <paco@voltanet.io>